requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deep-translator==1.9.0
fastapi==0.102.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langdetect==1.0.9
MarkupSafe==3.0.1
nltk==3.7
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.12
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
starlette==0.27.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
